#6 푸터로 로그인, 로그아웃 옮김.
