+++ app/login/page.tsx
@@ -26,6 +26,7 @@ export default function Login() {
         headers: { Authorization: `Bearer ${idToken}` },
       });
       router.push('/');
+      router.refresh();
     } catch (e) {
       setError((e as Error).message);
     }
@@ -36,7 +37,7 @@ export default function Login() {
       <div className="w-full bg-white rounded-lg shadow dark:border md:mt-0 sm:max-w-md xl:p-0 dark:bg-gray-800 dark:border-gray-700">
         <div className="p-6 space-y-4 md:space-y-6 sm:p-8">
           <h1 className="text-xl font-bold leading-tight tracking-tight text-gray-900 md:text-2xl dark:text-white">
-            Speak thy secret word!
+            로그인
           </h1>
           <form
             onSubmit={handleSubmit}
@@ -48,7 +49,7 @@ export default function Login() {
                 htmlFor="email"
                 className="block mb-2 text-sm font-medium text-gray-900 dark:text-white"
               >
-                Your email
+                아이디
               </label>
               <input
                 type="email"
@@ -57,7 +58,7 @@ export default function Login() {
                 onChange={(e) => setEmail(e.target.value)}
                 id="email"
                 className="bg-gray-50 border border-gray-300 text-gray-900 sm:text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
-                placeholder="name@company.com"
+                placeholder="email@naver.com"
                 required
               />
             </div>
@@ -66,7 +67,7 @@ export default function Login() {
                 htmlFor="password"
                 className="block mb-2 text-sm font-medium text-gray-900 dark:text-white"
               >
-                Password
+                비밀번호
               </label>
               <input
                 type="password"
@@ -91,15 +92,15 @@ export default function Login() {
               type="submit"
               className="w-full text-white bg-gray-600 hover:bg-gray-700 focus:ring-4 focus:outline-none focus:ring-primary-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-gray-600 dark:hover:bg-gray-700 dark:focus:ring-primary-800"
             >
-              Enter
+              로그인하기
             </button>
             <p className="text-sm font-light text-gray-500 dark:text-gray-400">
-              Don&apos;t have an account?
+              계정이 없으신가요?
               <Link
                 href="/register"
-                className="font-medium text-gray-600 hover:underline dark:text-gray-500"
+                className="ml-4 font-medium text-gray-600 hover:underline dark:text-gray-500"
               >
-                Register here
+                회원가입하기
               </Link>
             </p>
           </form>
